Reorganized file structure and fixed SVGs not loading
