+++ src/@types/bridge.d.ts
@@ -1,4 +1,4 @@
-import { API } from './index.js'
+import { API } from '../preload/index.js'
 
 declare global {
     interface Window { api: typeof API }
