+++ src/common/constants.ts
@@ -1,0 +1,6 @@
+module Constants {
+    export const MINUTES_TO_MS = 60000
+    export const MS_TO_MINUTES = 1 / MINUTES_TO_MS
+}
+
+export { Constants }
