+++ src/common/inputForm.ts
@@ -1,0 +1,173 @@
+import "../common/htmlElement"
+import "../common/htmlFormElement"
+import "../common/date"
+
+class InputForm {
+    formElement: HTMLFormElement
+    inputs: Map<String, HTMLInputElement | HTMLTextAreaElement | HTMLButtonElement>
+
+    constructor(formClass: string, onSubmit: (e: Event) => boolean, onReset: (e: Event) => boolean) {
+        this.inputs = new Map()
+        this.formElement = <HTMLFormElement>document.getElementsByClassName(formClass)[0]
+
+        this.formElement.addEventListener('submit', e => onSubmit(e))
+        this.formElement.addEventListener('reset', e => onReset(e))
+
+        const inputElements: Array<HTMLInputElement | HTMLTextAreaElement | HTMLButtonElement>
+             = Array.from(this.formElement.querySelectorAll('input,button,textarea'))
+
+        inputElements.forEach(e => {
+            const id = e.getAttribute('id');
+            const type = e.getAttribute('type')
+
+            if(id == null)
+                return
+
+            // Handle the error message
+            if((e instanceof HTMLInputElement || e instanceof HTMLTextAreaElement)) {
+                const errorMessage = document.createElement('p')
+                errorMessage.classList.add('error-message')
+
+                const updateValidationMessage = () => { errorMessage.innerHTML = e.validationMessage }
+
+                e.insertAdjacentElement("afterend", errorMessage)
+
+                e.onkeyup = updateValidationMessage
+                e.onmousedown = updateValidationMessage
+                updateValidationMessage()
+
+                e.oninvalid = () => {
+                    e.setDirty(true)
+                    updateValidationMessage()
+                }
+            }
+
+            // Add unit selection dropdowns
+            const useUnits = e.getAttribute('use-units')
+            if(useUnits) {
+                switch(useUnits) {
+                    case 'time':
+                        const units = document.createElement('span')
+                        units.id = `${id}-units`
+                        units.classList.add('units')
+                        e.insertAdjacentElement("afterend", units)
+
+                        units.innerHTML = 'minutes'
+                        break;
+                }
+                
+            }
+
+            switch(type) {
+                case 'checkbox':
+                    const toggles = e.getAttribute('toggles')
+                    if(toggles == null || !(e instanceof HTMLInputElement))
+                        break
+
+                    e.onchange = () => { 
+                        const input = this.inputs.get(toggles)
+                        if(input == null)
+                            return;
+                        
+                        input.disabled = !e.checked
+                    }
+
+                    break
+                default:
+                    break
+            }
+
+            e.onkeydown = () => e.setDirty(true)
+            e.onmousedown = () => e.setDirty(true)
+
+            this.inputs.set(id, e)
+        })
+    }
+
+    setValue(input: string, value: any) {
+        const element: any = this.getInputElement(input)
+
+        if(element == null)
+            return
+
+        if(!element.disabled)
+            element.value = value.toString();
+        else
+            element.value = ''
+    }
+
+    getValue(input: string, checkActive: boolean = false) {
+        if(checkActive && !this.activeAndFilled(input))
+            return ''
+
+        return this.getInputElement(input)?.value || ''
+    }
+
+    getValueAsNumber(input: string, checkActive: boolean = false) {
+        if(checkActive && !this.activeAndFilled(input))
+            return ''
+
+        const element = this.getInputElement(input)
+        if(!element || !(element instanceof HTMLInputElement))
+            return ''
+
+        return element.valueAsNumber
+    }
+
+    hasValue(input: string) {
+        return (this.inputs.get(input)?.value?.length || 0) > 0
+    }
+
+    activeAndFilled(input: string): boolean {
+        const inputElement = this.getInputElement(input)
+        if(inputElement == null)
+            return false;
+
+        return !inputElement.disabled && inputElement.value.length > 0
+    }
+
+    setChecked(input: String, checked: boolean) {
+        const element = this.inputs.get(input)
+        if(!element || !(element instanceof HTMLInputElement) || element.getAttribute('type') !== 'checkbox')
+            return
+        
+        element.checked = checked
+        element.dispatchEvent(new Event('change'))
+    }
+
+    getInputElement(input: String): HTMLInputElement | HTMLTextAreaElement | HTMLButtonElement | undefined  {
+        return this.inputs.get(input) || undefined
+    }
+
+    setFromJson(json: string): void {
+        const camelCaseRegex = /.([a-z])+/g
+    
+        // Set all the fields
+        const obj = JSON.parse(json)
+        for(let key in obj) {
+            const id = key.match(camelCaseRegex)?.flatMap(s => s.toLowerCase()).join('-') || ''
+            const element = <HTMLInputElement>document.getElementById(id)
+
+            if(element == null)
+                continue
+
+            if(element as HTMLInputElement | HTMLTextAreaElement) {}
+                element.value = obj[key]
+        }
+
+        // Set the toggle checkboxes
+        Array.from(this.inputs.values()).forEach(input => {
+            const type = input.getAttribute('type') 
+            if(type !== 'checkbox')
+                return
+            
+            const toggles = input.getAttribute('toggles')
+            if(toggles == null)
+                return
+
+            this.setChecked(input.id, this.hasValue(toggles))
+        })
+    }
+}
+
+export { InputForm }
