+++ src/renderer/reminder.html
@@ -92,6 +92,6 @@
         </section>>
     </div>
 
-    <script type="module" src="./src/renderer.ts"></script>
+    <script type="module" src="./src/reminder.ts"></script>
 </body>
 </html>
