+++ tsconfig.web.json
@@ -1,6 +1,6 @@
 {
     "extends": "@electron-toolkit/tsconfig/tsconfig.web.json",
-    "include": ["src/renderer/**/*.ts", "src/preload/*.d.ts", "src/preload/bridge.d.ts"],
+    "include": ["src/renderer/**/*.ts", "src/preload/*.d.ts", "src/@types/bridge.d.ts", "src/common/*.ts", "src/@types/svg.d.ts"],
     "compilerOptions": {
       "composite": true
     }
